requirements.txt
flask
flask-jwt-extended
flask-cors
python-dotenv
sqlalchemy
faiss-cpu
pdfplumber
tiktoken
openai
numpy